feat(usageguard): unify the usage scale and fix the self-inflicted throttle
The guard consults provider usage before every agent launch. The sources
were already the right ones, but three things around them were wrong.

Percentages were a convention held in a comment. Three readers each
decided locally how to interpret their provider's number, and only one
of them validated anything. Percent now makes the canonical form - share
of a window consumed, 0 to 100 - a real type, reached only through Used()
or Remaining(). The compiler enforces the boundary. A value off the
declared scale is rejected rather than clamped, because clamping keeps
the guard running against a number it cannot interpret. The scale is
declared per provider rather than sniffed: 0.04 is equally valid as
"0.04% used" and as a fraction meaning "4% used", so no single value can
reveal which convention produced it.

Readings now carry when they were measured, which is not when they were
read. This matters because the providers differ: Codex writes a fresh
percentage to its session log every turn, while Claude's local cache was
found 25.9 hours behind, reporting the weekly window at 100% when it was
actually at 0%. Acting on that would have blocked every Claude agent
indefinitely. A measurement past max_reading_age_seconds is reported as
stale and excluded, exactly like a window the provider never sent.

Finally the throttle. GET /api/oauth/usage rate-limits after roughly six
rapid calls, and Invalidate ran after every agent execution, so a run
made one call per invocation and walked into its own 429 - measured at
8 runs costing 8 calls. The 429 then escalated to Claude's interactive
/usage panel, which is backed by the same account and endpoint, so it
spent the full 25s timeout before failing identically and switching
provider while the subscription sat nearly untouched. A 429 is now
recognised and does not escalate; the guard retains its last good
reading and reuses it, subject to the same age limit; and Invalidate
expires the TTL without discarding that fallback. Codex additionally
gains a local session reader chained behind its app-server, matching how
existing monitors read it.

Verified by mutation - each of these fails a distinct test: widening the
age limit, discarding the retained reading, treating 429 as an ordinary
error, and accepting values off the declared scale.

+++ internal/invoke/usage_guard_live_probe_test.go
@@ -52,9 +52,9 @@ func liveGuard(t *testing.T, limit float64) *usageguard.Guard {
 // budget, so the role must skip Claude and complete on Codex.
 func TestLiveUsageGuardSwitchesWhenThresholdJustBelowCurrent(t *testing.T) {
 	live := liveClaudeFiveHour(t)
-	limit := live.UsedPercent - 0.5
+	limit := float64(live.UsedPercent) - 0.5
 	if limit <= 0 {
-		limit = math.Nextafter(live.UsedPercent, 0)
+		limit = math.Nextafter(float64(live.UsedPercent), 0)
 	}
 	if limit <= 0 {
 		t.Skip("live 5h usage is 0%; cannot pin a threshold below it")
@@ -92,7 +92,7 @@ func TestLiveUsageGuardBlocksAtExactlyCurrentUsage(t *testing.T) {
 
 	r := &scriptedContractRunner{}
 	inv, _ := contractRetryInvoker(t, r, []config.AgentSpec{{Name: "claude-only", Provider: "claude"}}, nil)
-	inv.UsageGuard = liveGuard(t, live.UsedPercent)
+	inv.UsageGuard = liveGuard(t, float64(live.UsedPercent))
 
 	_, err := Raw(context.Background(), inv, "coder", RoleCall{}, RunContext{TaskID: "T-live-exact"}, statusOKValidate)
 	if !errors.Is(err, ErrUsageThreshold) {
@@ -108,7 +108,7 @@ func TestLiveUsageGuardBlocksAtExactlyCurrentUsage(t *testing.T) {
 // unconditionally.
 func TestLiveUsageGuardAllowsWhenThresholdJustAboveCurrent(t *testing.T) {
 	live := liveClaudeFiveHour(t)
-	limit := live.UsedPercent + 0.5
+	limit := float64(live.UsedPercent) + 0.5
 	if limit > 100 {
 		t.Skip("live 5h usage is at 100%; cannot pin a threshold above it")
 	}
